pyproject.toml
[tool.poetry]
name = "nestor-car-webapp"
version = "0.1.0"
description = ""
authors = ["Ana Nestor <anapfnestor@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
django-phonenumber-field = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
